+++ src/main/java/net/geforcemods/securitycraft/api/IEMPAffected.java
@@ -1,0 +1,15 @@
+package net.geforcemods.securitycraft.api;
+
+public interface IEMPAffected {
+	public default void shutDown() {
+		setShutDown(true);
+	}
+
+	public default void reactivate() {
+		setShutDown(false);
+	}
+
+	public boolean isShutDown();
+
+	public void setShutDown(boolean shutDown);
+}