+++ src/main/java/net/geforcemods/securitycraft/compat/icbmclassic/ICBMClassicEMPCompat.java
@@ -1,8 +1,9 @@
 package net.geforcemods.securitycraft.compat.icbmclassic;
 
 import icbm.classic.api.events.EmpEvent;
+import net.geforcemods.securitycraft.api.IEMPAffected;
 import net.geforcemods.securitycraft.entity.camera.EntitySecurityCamera;
-import net.geforcemods.securitycraft.tileentity.IEMPAffected;
+import net.minecraft.entity.Entity;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
@@ -17,9 +18,14 @@ public class ICBMClassicEMPCompat {
 
 	@SubscribeEvent
 	public void onEmpEntityPost(EmpEvent.EntityPost event) {
-		if (event.target instanceof EntitySecurityCamera) {
-			event.target.removePassengers();
-			event.target.setDead();
+		Entity entity = event.target;
+
+		if (entity instanceof IEMPAffected)
+			((IEMPAffected) entity).shutDown();
+
+		if (entity instanceof EntitySecurityCamera) {
+			entity.removePassengers();
+			entity.setDead();
 		}
 	}
 }