+++ src/main/java/net/geforcemods/securitycraft/tileentity/TileEntitySonicSecuritySystem.java
@@ -9,6 +9,7 @@ import java.util.Set;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.api.CustomizableSCTE;
 import net.geforcemods.securitycraft.api.ILockable;
+import net.geforcemods.securitycraft.api.ITEEMPAffected;
 import net.geforcemods.securitycraft.api.Option;
 import net.geforcemods.securitycraft.api.Option.OptionInt;
 import net.geforcemods.securitycraft.blocks.BlockSonicSecuritySystem;
@@ -27,7 +28,7 @@ import net.minecraft.util.SoundCategory;
 import net.minecraft.util.math.BlockPos;
 import net.minecraftforge.common.util.Constants;
 
-public class TileEntitySonicSecuritySystem extends CustomizableSCTE implements ITickable, IEMPAffected {
+public class TileEntitySonicSecuritySystem extends CustomizableSCTE implements ITickable, ITEEMPAffected {
 	/** The delay between each ping sound in ticks */
 	private static final int PING_DELAY = 100;
 	/**